fix(driver/bpf): sys_open_e verifier issue on aarch64

+++ driver/ppm_flag_helpers.h
@@ -25,7 +25,7 @@ or GPL2.txt for full copies of the license.
 #define PPM_MS_MGC_MSK 0xffff0000
 #define PPM_MS_MGC_VAL 0xC0ED0000
 
-static __always_inline uint32_t open_flags_to_scap(unsigned long flags)
+static __always_inline uint32_t open_flags_to_scap(uint32_t flags)
 {
 	uint32_t res = 0;
